FLINK-6041 Move StreamingFunctionUtils to 'org.apache.flink.streaming.util'
